Continue to impl. of E2E encryption

+++ include/e2e/ParticipantKeyHandler.h
@@ -34,10 +34,10 @@ public:
                           const std::shared_ptr<Bricks::Logger>& logger = {});
     virtual ~ParticipantKeyHandler();
     std::shared_ptr<ParticipantKeyHandler> clone() const;
-    virtual std::vector<uint8_t> ratchetKey(const std::optional<size_t>& keyIndex = {});
-    virtual std::shared_ptr<KeySet> keySet(const std::optional<size_t>& keyIndex = {}) const;
+    virtual std::vector<uint8_t> ratchetKey(const std::optional<uint8_t>& keyIndex = {});
+    virtual std::shared_ptr<KeySet> keySet(const std::optional<uint8_t>& keyIndex = {}) const;
     virtual void setKey(std::vector<uint8_t> password,
-                        const std::optional<size_t>& keyIndex = {});
+                        const std::optional<uint8_t>& keyIndex = {});
     std::vector<uint8_t> ratchetKeyMaterial(const std::vector<uint8_t>& currentMaterial) const;
     std::shared_ptr<KeySet> deriveKeys(const std::vector<uint8_t>& ratchetSalt,
                                        std::vector<uint8_t> password,
@@ -45,7 +45,7 @@ public:
     bool hasValidKey() const;
     void setHasValidKey();
     void setKeyFromMaterial(std::vector<uint8_t> password,
-                            const std::optional<size_t>& keyIndex = {});
+                            const std::optional<uint8_t>& keyIndex = {});
     bool decryptionFailure();
 private:
     ParticipantKeyHandler(std::unique_ptr<Impl> impl);
