+++ src/webrtc/media/e2e/DefaultKeyProvider.h
@@ -31,17 +31,17 @@ public:
                        const std::shared_ptr<Bricks::Logger>& logger = {});
     ~DefaultKeyProvider() final;
     // impl. of KeyProvider
-    bool setSharedKey(std::vector<uint8_t> key, const std::optional<size_t>& keyIndex = {}) final;
+    bool setSharedKey(std::vector<uint8_t> key, const std::optional<uint8_t>& keyIndex = {}) final;
     std::shared_ptr<ParticipantKeyHandler> sharedKey(const std::string& participantId) final;
-    std::vector<uint8_t> ratchetSharedKey(const std::optional<size_t>& keyIndex = {}) final;
-    std::vector<uint8_t> exportSharedKey(const std::optional<size_t>& keyIndex = {}) const final;
+    std::vector<uint8_t> ratchetSharedKey(const std::optional<uint8_t>& keyIndex = {}) final;
+    std::vector<uint8_t> exportSharedKey(const std::optional<uint8_t>& keyIndex = {}) const final;
     bool setKey(const std::string& participantId, std::vector<uint8_t> key,
-                const std::optional<size_t>& keyIndex = {}) final;
+                const std::optional<uint8_t>& keyIndex = {}) final;
     std::shared_ptr<ParticipantKeyHandler> key(const std::string& participantId) const final;
     std::vector<uint8_t> ratchetKey(const std::string& participantId,
-                                    const std::optional<size_t>& keyIndex = {}) const final;
+                                    const std::optional<uint8_t>& keyIndex = {}) const final;
     std::vector<uint8_t> exportKey(const std::string& participantId,
-                                   const std::optional<size_t>& keyIndex = {}) const final;
+                                   const std::optional<uint8_t>& keyIndex = {}) const final;
     void setSifTrailer(std::vector<uint8_t> trailer) final;
     std::vector<uint8_t> sifTrailer() const final;
     const KeyProviderOptions& options() const final;
